@@ -2,33 +2,26 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package io.flutter.plugins.geofencing
+package com.omnys.geofencing.flutter_geofencing_omnys
 
 import android.content.Context
 import android.content.Intent
-import android.os.IBinder
-import android.os.PowerManager
 import android.os.Handler
+import android.os.Parcelable
 import android.util.Log
 import androidx.core.app.JobIntentService
+import io.flutter.embedding.engine.FlutterEngine
+import io.flutter.embedding.engine.dart.DartExecutor.DartCallback
+import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.PluginRegistry.PluginRegistrantCallback
 import io.flutter.view.FlutterCallbackInformation
 import io.flutter.view.FlutterMain
-import io.flutter.view.FlutterNativeView
-import io.flutter.view.FlutterRunArguments
+import org.altbeacon.beacon.Region
 import java.util.ArrayDeque
-import java.util.concurrent.atomic.AtomicBoolean
 import java.util.UUID
-
-import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.embedding.engine.dart.DartExecutor
-import io.flutter.embedding.engine.dart.DartExecutor.DartCallback
-
-import com.google.android.gms.location.GeofencingEvent
+import java.util.concurrent.atomic.AtomicBoolean
 
 class GeofencingService : MethodCallHandler, JobIntentService() {
     private val queue = ArrayDeque<List<Any>>()
@@ -38,10 +31,13 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
     companion object {
         @JvmStatic
         private val TAG = "GeofencingService"
+
         @JvmStatic
         private val JOB_ID = UUID.randomUUID().mostSignificantBits.toInt()
+
         @JvmStatic
         private var sBackgroundFlutterEngine: FlutterEngine? = null
+
         @JvmStatic
         private val sServiceStarted = AtomicBoolean(false)
 
@@ -57,6 +53,24 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
         fun setPluginRegistrant(callback: PluginRegistrantCallback) {
             sPluginRegistrantCallback = callback
         }
+
+        @JvmStatic
+        fun createIntent(context: Context, region: Region, eventType: EventType): Intent {
+            val intent = Intent(context, GeofencingService::class.java)
+            intent.putExtra("region", region as Parcelable)
+            intent.putExtra("eventType", eventType)
+            return intent
+        }
+
+        @JvmStatic
+        fun parseIntentRegion(intent: Intent): Region? {
+            return intent.getParcelableExtra("region")
+        }
+
+        @JvmStatic
+        fun parseIntentEventType(intent: Intent): EventType? {
+            return intent.getSerializableExtra("eventType") as EventType?
+        }
     }
 
     private fun startGeofencingService(context: Context) {
@@ -66,15 +80,17 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
                 sBackgroundFlutterEngine = FlutterEngine(context)
 
                 val callbackHandle = context.getSharedPreferences(
-                        GeofencingPlugin.SHARED_PREFERENCES_KEY,
-                        Context.MODE_PRIVATE)
-                        .getLong(GeofencingPlugin.CALLBACK_DISPATCHER_HANDLE_KEY, 0)
+                    FlutterGeofencingOmnysPlugin.SHARED_PREFERENCES_KEY,
+                    Context.MODE_PRIVATE
+                )
+                    .getLong(FlutterGeofencingOmnysPlugin.CALLBACK_DISPATCHER_HANDLE_KEY, 0)
                 if (callbackHandle == 0L) {
                     Log.e(TAG, "Fatal: no callback registered")
                     return
                 }
 
-                val callbackInfo = FlutterCallbackInformation.lookupCallbackInformation(callbackHandle)
+                val callbackInfo =
+                    FlutterCallbackInformation.lookupCallbackInformation(callbackHandle)
                 if (callbackInfo == null) {
                     Log.e(TAG, "Fatal: failed to find callback")
                     return
@@ -82,21 +98,23 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
                 Log.i(TAG, "Starting GeofencingService...")
 
                 val args = DartCallback(
-                    context.getAssets(),
+                    context.assets,
                     FlutterMain.findAppBundlePath(context)!!,
                     callbackInfo
                 )
-                sBackgroundFlutterEngine!!.getDartExecutor().executeDartCallback(args)
+                sBackgroundFlutterEngine!!.dartExecutor.executeDartCallback(args)
                 IsolateHolderService.setBackgroundFlutterEngine(sBackgroundFlutterEngine)
             }
         }
-        mBackgroundChannel = MethodChannel(sBackgroundFlutterEngine!!.getDartExecutor().getBinaryMessenger(),
-                "plugins.flutter.io/geofencing_plugin_background")
+        mBackgroundChannel = MethodChannel(
+            sBackgroundFlutterEngine!!.dartExecutor.binaryMessenger,
+            "plugins.flutter.io/geofencing_plugin_background"
+        )
         mBackgroundChannel.setMethodCallHandler(this)
     }
 
-   override fun onMethodCall(call: MethodCall, result: Result) {
-       when(call.method) {
+    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        when (call.method) {
             "GeofencingService.initialized" -> {
                 synchronized(sServiceStarted) {
                     while (!queue.isEmpty()) {
@@ -105,14 +123,17 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
                     sServiceStarted.set(true)
                 }
             }
+
             "GeofencingService.promoteToForeground" -> {
                 mContext.startForegroundService(Intent(mContext, IsolateHolderService::class.java))
             }
+
             "GeofencingService.demoteToBackground" -> {
                 val intent = Intent(mContext, IsolateHolderService::class.java)
-                intent.setAction(IsolateHolderService.ACTION_SHUTDOWN)
+                intent.action = IsolateHolderService.ACTION_SHUTDOWN
                 mContext.startForegroundService(intent)
             }
+
             else -> result.notImplemented()
         }
         result.success(null)
@@ -124,29 +145,22 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
     }
 
     override fun onHandleWork(intent: Intent) {
-        val callbackHandle = intent.getLongExtra(GeofencingPlugin.CALLBACK_HANDLE_KEY, 0)
-        val geofencingEvent = GeofencingEvent.fromIntent(intent)
-        if (geofencingEvent.hasError()) {
-            Log.e(TAG, "Geofencing error: ${geofencingEvent.errorCode}")
+        val region = parseIntentRegion(intent);
+        val eventType = parseIntentEventType(intent)
+        if (region == null || eventType == null) {
+            Log.e(TAG, "Geofencing error: region $region, eventType $eventType")
             return
         }
 
-        // Get the transition type.
-        val geofenceTransition = geofencingEvent.geofenceTransition
-
-        // Get the geofences that were triggered. A single event can trigger
-        // multiple geofences.
-        val triggeringGeofences = geofencingEvent.triggeringGeofences.map {
-            it.requestId
+        val geofenceFromCache =
+            FlutterGeofencingOmnysPlugin.getGeofenceFromCache(applicationContext, region.uniqueId)
+        if (geofenceFromCache == null) {
+            Log.e(TAG, "Geofencing error: no geofence found in case for ${region.uniqueId}");
+            return
         }
+        val callbackHandle = geofenceFromCache.get(0) as Long
 
-        val location = geofencingEvent.triggeringLocation
-        val locationList = listOf(location.latitude,
-                location.longitude)
-        val geofenceUpdateList = listOf(callbackHandle,
-                triggeringGeofences,
-                locationList,
-                geofenceTransition)
+        val geofenceUpdateList = listOf(callbackHandle, listOf(region.uniqueId), eventType.dartValue)
 
         synchronized(sServiceStarted) {
             if (!sServiceStarted.get()) {
@@ -154,7 +168,12 @@ class GeofencingService : MethodCallHandler, JobIntentService() {
                 queue.add(geofenceUpdateList)
             } else {
                 // Callback method name is intentionally left blank.
-                Handler(mContext.mainLooper).post { mBackgroundChannel.invokeMethod("", geofenceUpdateList) }
+                Handler(mContext.mainLooper).post {
+                    mBackgroundChannel.invokeMethod(
+                        "",
+                        geofenceUpdateList
+                    )
+                }
             }
         }
     }
