@@ -6,9 +6,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-
-import 'package:geofencing/src/geofencing.dart';
-import 'package:geofencing/src/location.dart';
+import 'package:flutter_geofencing_omnys/src/geofencing.dart';
 
 void callbackDispatcher() {
   const MethodChannel _backgroundChannel =
@@ -17,19 +15,12 @@ void callbackDispatcher() {
 
   _backgroundChannel.setMethodCallHandler((MethodCall call) async {
     final List<dynamic> args = call.arguments;
-    final Function callback = PluginUtilities.getCallbackFromHandle(
+    final Function? callback = PluginUtilities.getCallbackFromHandle(
         CallbackHandle.fromRawHandle(args[0]));
     assert(callback != null);
     final List<String> triggeringGeofences = args[1].cast<String>();
-    final List<double> locationList = <double>[];
-    // 0.0 becomes 0 somewhere during the method call, resulting in wrong
-    // runtime type (int instead of double). This is a simple way to get
-    // around casting in another complicated manner.
-    args[2]
-        .forEach((dynamic e) => locationList.add(double.parse(e.toString())));
-    final Location triggeringLocation = locationFromList(locationList);
-    final GeofenceEvent event = intToGeofenceEvent(args[3]);
-    callback(triggeringGeofences, triggeringLocation, event);
+    final GeofenceEvent event = intToGeofenceEvent(args[2]);
+    callback!(triggeringGeofences, event);
   });
   _backgroundChannel.invokeMethod('GeofencingService.initialized');
 }
